requirements.txt
streamlit==1.31.0
requests==2.31.0
pandas>=2.2.0
google-generativeai>=0.8.0
google-genai>=1.47.0
python-dotenv>=1.0.0
plotly>=5.18.0
praw>=7.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
